added a command line option to notify the archive creation
When passing --notify as command line argument, after the creation of an
archive, a dialog will be displayed to confirm the successful completion of
the operation allowing the user to open the archive if he wants to.  If the
progress dialog has lost the focus the notification will be shown using the
desktop notification system (if compiled with libnotify).

This option is used when creating archives from the Nautilus context menu
command.