fr-window: Improve XDS dnd support
Improve file-roller XDS dnd support by sending the 'success'
response only after the file(s) has already been extracted, so
the target dnd application can operate on it (eg. gedit may
want to open it, nautilus may want to reveal it).

The XDS spec states that when sending the success ('S') response, the
file should have already been written to disk, but assumed this would
be done in a synchronous (and thus blocking) operation, asynchronous
(callback based) apis were not common when the XDS spec was done so
that's probably the reason.

So as file-roller does, in a right manner, its file operations asynchronously,
thats probably why it didn't care to send the success response at the right time.

This patch workarounds the XDS spec limitation and makes file-roller to wait
for the async operations to complete before sending the success response.

It does so, without blocking the main ui thread, by creating and using a GMainLoop
the same way as it's done by the gtk_dialog_run() call, which shows a dialog and
waits for its response without blocking the ui.

Fixes bug 710546