Fix Add Files dialog button sensitivity
When the Add Files dialog has previously been opened and there are
selected files, the selection is restored for subsequent views of the
dialog. However, the sensitivity of the Add button is only updated when
manually selecting a file, as the dialog only updates the sensitivity
in response to the GtkCellRendererToggle for the selected state being
changed. When setting a list of selected files on the dialog, as done
when restoring the list of selected files on subsequent views of the
dialog, the Add button is only ever set to be insensitive.

Fix this by updating the Add button sensitivity after the list of files
has been loaded.

https://bugzilla.gnome.org/show_bug.cgi?id=738177