Fix renaming files in a password-protected archive
When opening a password-protected archive, and renaming a file within,
the password dialog takes user input, sets the password for the archive
and then restarts the rename action. As part of doing so, it frees the
existing action and replaces it with a duplicate. However, pointers to
fields inside the original (now freed) action struct are used to
provide the old and new filenames, among other arguments. This leads to
the file being deleted (although accessing freed memory is undefined,
so other outcomes are possible).

Fix the rename operation by always using the fields from the new
struct, which exists for the lifetime of the rename action.

https://bugzilla.redhat.com/show_bug.cgi?id=1233853
https://bugzilla.gnome.org/show_bug.cgi?id=766227