Revert "Remove nautilus extension"
This reverts commit da09ee41ca7c9b63082cf2a35ae19701c34adca7.

Some of the main compressed archive formats are not still supported by
libarchive, for example, RAR5.

This is a major issue for the Nautilus builtin compression handling, so
for now, and until libarchive adds support for RAR5, let's enable the
nautilus extension of file-roller for the "Extract here" menu item.

See https://github.com/libarchive/libarchive/issues/373 for libarchive
support of RAR5.

https://bugzilla.gnome.org/show_bug.cgi?id=772765