nautilus-fileroller: remove compress support
Recently we reverted the removal of the Nautilus extension due to
libarchive not having RAR5 support. The commit in question is 366a5147b.

However, the compress handling is still done fine on Nautilus side, so
that part of the Nautilus extension shouldn't be present.

This commit removes the "compress" menu item and its associated code in
the Nautilus extension.

https://bugzilla.gnome.org/show_bug.cgi?id=779706