dlg-package-installer.c: Remove X11 dependency
When a user opens an archive of a type that no installed package can open
a dialog pops up with the option to search for a suitable application
in GNOME Software.

This functionality was provided by org.freedesktop.PackageKit.Modify's
InstallPackageNames method that required an X11 window id.

The fix consists of migrating to the new org.freedesktop.PackageKit.Modify2
D-Bus API that handles this use-case without needing an X11 window id and
instead using a reverse domain name application id.

Closes #78