Add files for Nix development environment
Nix [1] is purely functional package manager allowing to easily
install dependencies in a reproducible manner.
A developer can run `nix-shell` in the project directory
and they will be dropped into a shell with all the necessary
build dependencies.

The dependencies are pinned so that same version can be
used on CI and developers’ machines.
One can run `nix-shell --run 'niv update'` to update to latest version.

[1]: https://nixos.org/