build: Find python for helper script explicitly
If Python is not present (e.g. when building in a sandbox),
Meson will fail in the middle of a build:

    /usr/bin/env: ‘python3’: No such file or directory

Let’s check for `python3` using `find_program` so that
the issue can be detected in the configure step.

It would be even nicer to get rid of the Python script
in favour of using Meson’s filesystem module to read
the supported-mime-types file in Meson itself but
the `fs.read` function will currently trigger full
Meson re-configuration when a change in the file is detected,
instead of just invalidation of the affected targets
like `configure_file` does.