introspection: Allow building without schemas installed and outside X server
Previously, we handled the `--introspection-dump` CLI flag inside GApplication
but that initialized bunch of classes that bind GSettings in their constructors
so they need access to schemas, which are not yet available when building
in a sandbox. The option parser itself adds GTK options, which triggers
GTK to check for display server.

Let’s move handling the `--introspection-dump` to the beginning of the main
program entry point before GApplication is even created to prevent this.