command-7z: Fix detection of unrar potential support
The potential capabilities introduced in
https://gitlab.gnome.org/GNOME/file-roller/-/commit/e7fe13ab35956fe1c68c4e3fa35b9e3e2f155568
which can be gained by installing extra software should clearly be
a superset of the current capabilities.

But this is not the case when a `FrArchive` subclass checks
for the availability of a program to “give way” to another
`FrArchive` subclass that supports more features for given format.
This is caused by the fact that when scanning for a potential
capabilities, the `check_command` argument is `TRUE` and `_g_program_is_available`
function will pretend that the program is available.
This will cause the `FrArchive` instance deny its potential capabilities
because it will yield to a different one.
https://gitlab.gnome.org/GNOME/file-roller/-/commit/160fd2dab3977fa9fdf727cb74e6a23b8fc9d20b