NEWS: Update list of changes
Let’s keep the NEWS up to date when we can to reduce the amount
of work needed to be done when releasing.
This was suggested in https://discourse.gnome.org/t/changing-the-way-gnome-modules-are-released/10443,
though it does not actually applies to us, since File Roller is not in GNOME Core any more.