command-rar: Fix ask password for encrypted RARs
fr_command_rar_handle_error () searched for certain strings in the
command output to detect the password was wrong or missing. This didn't
work correctly from at least version 5.50 of rar/unrar and made it
impossible to extract files from encrypted RARs or open RARs with an
encrypted header, because file-roller wouldn't ask for a password.

From at least version 5.50 of rar/unrar the commands give exit value 11
when the password was wrong or missing. fr_command_rar_handle_error ()
is changed to check for that exit value and now asks for a password
when that is needed.

Fixes #139.