lha: Fix crash, parse lha list output correctly
The output from the `lha` list archive command is eclectic and
inconsistent, and file-roller was actually crashing when opening certain
archives (eg. lha_amiga_212/level1.lzh in the Lhasa test suite). But
this also fixes the list parsing more generally. With this change,
file-roller now appears to successfully open all of the archive files in
the Lhasa test suite correctly without crashing.

The first field in lha's list output either contains Unix permissions or
an OS name in [brackets]. There was already hard-coded support for
[MS-DOS], [generic] and [unknown] but other OS names were not handled
properly. This is now fixed, with an approach that is mindful of the
fact that the OS name can contain spaces.

Empty value (whitespace) in the UID/GID column is also handled
correctly.  This was partially working because of the aforementioned
special-casing, but now is fixed more generally.